bug(UIKIT-1470): Реализована секция next, для компонентов с требованием react >= 18 версии
