+++ packages/components/pack.js
@@ -16,5 +16,11 @@ buildTs({
       module: './server/index.js',
       require: './node/server/index.js',
     },
+    './next': {
+      // Секция для компонентов, для которых необходим react >= 18 версии
+      vitest: './node/server/next.js',
+      module: './server/next.js',
+      require: './node/server/next.js',
+    },
   },
 });
