+++ packages/components/src/next.ts
@@ -1,0 +1,9 @@
+// Компоненты, для работы которых необходим react >= 18 версии
+
+export {
+  notify as notifyNext,
+  NotificationContainerNext,
+  NotificationStackContainerNext,
+  NOTIFY_POSITIONS,
+  type Notify as NotifyNext,
+} from './NotificationNext';
